# backend/openapischema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Document Analysis API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents/upload": {
      "post": {
        "tags": ["Documents"],
        "summary": "Upload documents to a default collection",
        "description": "Handles the bulk upload of one or more documents. If no collection is specified, documents are added to a default collection. Document processing (e.g., embedding creation) is handled asynchronously in a background task. Returns the details of the collection and the newly created documents.",
        "operationId": "default_document_upload_api_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_default_document_upload_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/collections/{collection_id}/documents/upload": {
      "post": {
        "tags": ["Documents"],
        "summary": "Upload documents to a specific collection",
        "description": "Handles the bulk upload of one or more documents to a specified collection. Document processing (e.g., embedding creation) is handled asynchronously in a background task. Returns a list of the newly created document details.",
        "operationId": "upload_documents_to_collection_api_v1_documents_collections__collection_id__documents_upload_post",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_to_collection_api_v1_documents_collections__collection_id__documents_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentInDB"
                  },
                  "title": "Response Upload Documents To Collection Api V1 Documents Collections  Collection Id  Documents Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/collections/{collection_id}/documents": {
      "get": {
        "tags": ["Documents"],
        "summary": "Retrieve all documents in a collection",
        "description": "Fetches a list of all documents belonging to a specific collection, identified by its ID. Returns an empty list if the collection has no documents.",
        "operationId": "read_documents_in_collection_api_v1_documents_collections__collection_id__documents_get",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentInDB"
                  },
                  "title": "Response Read Documents In Collection Api V1 Documents Collections  Collection Id  Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": ["Documents"],
        "summary": "Retrieve a single document by ID",
        "description": "Fetches a single document's details using its unique identifier. Returns a 404 error if the document is not found.",
        "operationId": "read_document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Documents"],
        "summary": "Delete a document",
        "description": "Deletes a document using its unique identifier. This action also triggers a background task to refresh the embeddings for the associated collection. Returns a 404 error if the document is not found.",
        "operationId": "delete_document_api_v1_documents__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_default_document_upload_api_v1_documents_upload_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files",
            "description": "List of files to upload."
          }
        },
        "type": "object",
        "required": ["files"],
        "title": "Body_default_document_upload_api_v1_documents_upload_post"
      },
...
      "Body_upload_documents_to_collection_api_v1_documents_collections__collection_id__documents_upload_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files",
            "description": "List of files to upload."
          }
        },
        "type": "object",
        "required": ["files"],
        "title": "Body_upload_documents_to_collection_api_v1_documents_collections__collection_id__documents_upload_post"
      },
...
      "CollectionInDB": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "The name of the collection."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "A brief description of the collection's purpose."
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "A list of tags for categorization."
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "The unique identifier for the collection."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat",
            "description": "The timestamp when the collection was created."
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat",
            "description": "The timestamp of the last update."
          },
          "total_docs": {
            "type": "integer",
            "title": "Total Docs",
            "description": "The total number of documents in the collection.",
            "default": 0
          },
          "latestInsightId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestinsightid",
            "description": "ID of the latest insight generated for this collection."
          },
          "latestPodcastId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestpodcastid",
            "description": "ID of the latest podcast generated for this collection."
          }
        },
        "type": "object",
        "required": ["id", "createdAt"],
        "title": "CollectionInDB",
        "description": "Pydantic model representing a collection as it is stored in the database.\n\nThis model is used for API responses and includes database-generated fields."
      },
...
      "DocumentInDB": {
        "properties": {
          "docTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctitle",
            "description": "A user-editable title for the document."
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "The unique identifier for the document."
          },
          "collectionId": {
            "type": "string",
            "title": "Collectionid",
            "description": "The ID of the collection this document belongs to."
          },
          "docName": {
            "type": "string",
            "title": "Docname",
            "description": "The original filename of the document."
          },
          "docSizeKB": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docsizekb",
            "description": "The size of the document in kilobytes."
          },
          "total_pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Pages",
            "description": "The total number of pages in the document."
          },
          "docType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctype",
            "description": "The MIME type of the document."
          },
          "docUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docurl",
            "description": "The URL where the document is stored."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat",
            "description": "The timestamp when the document was created."
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat",
            "description": "The timestamp of the last update."
          },
          "latestInsightId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestinsightid",
            "description": "ID of the latest insight generated for this document."
          },
          "latestPodcastId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestpodcastid",
            "description": "ID of the latest podcast generated for this document."
          },
          "isProcessed": {
            "$ref": "#/components/schemas/ProcessingStatus",
            "description": "The processing status of the document.",
            "default": "Pending"
          },
          "isEmbeddingCreated": {
            "$ref": "#/components/schemas/ProcessingStatus",
            "description": "The status of embedding creation for the document.",
            "default": "Pending"
          },
          "outline": {
            "items": {
              "$ref": "#/components/schemas/DocumentOutlineItemPydantic"
            },
            "type": "array",
            "title": "Outline",
            "description": "The structured outline of the document.",
            "default": []
          }
        },
        "type": "object",
        "required": ["id", "collectionId", "docName", "createdAt"],
        "title": "DocumentInDB",
        "example": {
          "collectionId": "collection_example_id_abc",
          "createdAt": "2025-01-01T12:00:00",
          "docName": "example_document.pdf",
          "docSizeKB": 1024,
          "docTitle": "Example Document Title",
          "docType": "application/pdf",
          "docUrl": "storage/uploads/collection_example_id_abc/example_document.pdf",
          "id": "doc_example_id_123",
          "isEmbeddingCreated": "Pending",
          "isProcessed": "Success",
          "outline": [],
          "total_pages": 5,
          "updatedAt": "2025-01-01T12:00:00"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProcessingStatus": {
        "type": "string",
        "enum": ["Pending", "Success", "Failed"],
        "title": "ProcessingStatus"
      },
      "UploadResponse": {
        "properties": {
          "collection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CollectionInDB"
              },
              {
                "type": "null"
              }
            ],
            "description": "The collection details associated with the uploaded documents."
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentInDB"
            },
            "type": "array",
            "title": "Documents",
            "description": "List of uploaded document details."
          }
        },
        "type": "object",
        "required": ["documents"],
        "title": "UploadResponse",
        "description": "Response model for document upload operations, including collection and document details."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}